+++ lib/src/templates/feature/repository_template.dart
@@ -8,16 +8,15 @@ String repositoryTemplate({
   final className = moduleName.singular.pascalCase;
   final entityFile = moduleName.singular.snakeCase;
   return '''
-import 'package:dartz/dartz.dart';
 import 'package:$projectName/features/$moduleName/domain/entities/$entityFile.dart';
 import 'package:$projectName/core/errors/failure.dart';
 
 abstract class ${className}Repository {
-  Future<Either<Failure, List<$className>>> getAll();
-  Future<Either<Failure, $className>> getById(String id);
-  Future<Either<Failure, $className>> create($className entity);
-  Future<Either<Failure, $className>> update($className entity);
-  Future<Either<Failure, void>> delete(String id);
+  Future<(List<$className>?, Failure?)> getAll();
+  Future<($className?, Failure?)> getById(String id);
+  Future<($className?, Failure?)> create($className entity);
+  Future<($className?, Failure?)> update($className entity);
+  Future<(bool, Failure?)> delete(String id);
 }
 ''';
 }